Streaming/requirements.txt
# AeroStream Streaming Requirements
psycopg2-binary>=2.9.9
requests>=2.31.0
streamlit>=1.29.0
pandas>=2.1.0
plotly>=5.18.0
